--- apps/gotask_ui/src/app/(portal)/goals/projectid/[projectId]/page.tsx
@@ -1,18 +1,0 @@
-"use client";
-import React from "react";
-import ModuleHeader from "@/app/component/header/moduleHeader";
-import { LOCALIZATION } from "@/app/common/constants/localization";
-import { useTranslations } from "next-intl";
-import ProjectGoalList from "@/app/(portal)/goals/projectid/[projectId]/components/projectGoal/projectGoalList";
-
-const Page = () => {
-  const transgoal = useTranslations(LOCALIZATION.TRANSITION.PROJECTGOAL);
-  return (
-    <>
-      <ModuleHeader name={transgoal("goal")} />
-      <ProjectGoalList />
-    </>
-  );
-};
-
-export default Page;
