+++ apps/gotask_ui/src/app/(portal)/project/viewProject/[projectId]/projectGoal/projectGoalForm.tsx
@@ -3,8 +3,8 @@ import FormField from "@/app/component/input/formField";
 import Grid from "@mui/material/Grid/Grid";
 import { useTranslations } from "next-intl";
 import { LOCALIZATION } from "@/app/common/constants/localization";
-import { ProjectGoalFormProps } from "../../interface/projectGoal";
 import { priorityOptions, statusOptions } from "@/app/common/constants/project";
+import { ProjectGoalFormProps } from "../interface/projectGoal";
 
 const ProjectGoalForm: React.FC<ProjectGoalFormProps> = ({ goalData, errors, setGoalData }) => {
   const transGoal = useTranslations(LOCALIZATION.TRANSITION.PROJECTGOAL);
